+++ src/subcommands/upgrade.rs
@@ -2,12 +2,14 @@ use crate::{CROSS, TICK, YELLOW_TICK};
 use anyhow::{bail, Result};
 use colored::Colorize;
 use ferinth::Ferinth;
+use fs_extra::file::{move_file, CopyOptions};
 use furse::Furse;
 use itertools::Itertools;
 use libium::config;
 use libium::upgrade;
 use octocrab::Octocrab;
-use tokio::fs::remove_file;
+use std::fs::read_dir;
+use tokio::fs::copy;
 
 struct Downloadable {
     filename: String,
@@ -183,41 +185,59 @@ pub async fn upgrade(
         );
     }
 
-    eprint!("\n{}", "Downloading Mod Files... ".bold());
-    for file in std::fs::read_dir(&profile.output_dir)? {
-        let file = file?;
-        let path = file.path();
-        if path.is_file() {
-            let mut index = None;
-            // If a file is already downloaded
-            if let Some(downloadable) = to_download
-                .iter()
-                .find_position(|thing| file.file_name().to_str().unwrap() == thing.filename)
-            {
-                index = Some(downloadable.0);
-            }
-            match index {
-                // Then don't download the file
-                Some(index) => {
-                    to_download.swap_remove(index);
-                },
-                // Or else delete the file
-                None => remove_file(path).await?,
+    let mut to_install = Vec::new();
+    if profile.output_dir.join("user").exists() {
+        for file in read_dir(&profile.output_dir.join("user"))? {
+            let file = file?;
+            let path = file.path();
+            if path.is_file() {
+                to_install.push((file.file_name(), path));
             }
         }
     }
-    match {
-        for downloadable in to_download {
-            let contents = reqwest::get(downloadable.download_url)
-                .await?
-                .bytes()
-                .await?;
-            upgrade::write_mod_file(profile, contents, &downloadable.filename).await?;
+
+    println!("\n{}\n", "Downloading Mod Files".bold());
+
+    for file in read_dir(&profile.output_dir)? {
+        let file = file?;
+        if file.file_type()?.is_file() {
+            let filename = file.file_name();
+            let filename = filename.to_str().unwrap();
+            if let Some((index, _)) = to_download
+                .iter()
+                .find_position(|thing| filename == thing.filename)
+            {
+                to_download.swap_remove(index);
+            } else if let Some((index, _)) =
+                to_install.iter().find_position(|thing| filename == thing.0)
+            {
+                to_install.swap_remove(index);
+            } else {
+                let _ = move_file(
+                    file.path(),
+                    profile.output_dir.join(".old").join(filename),
+                    &CopyOptions::new(),
+                );
+            }
         }
-        Ok::<(), anyhow::Error>(())
-    } {
-        Ok(_) => println!("{}", *TICK),
-        Err(_) => bail!("{}", CROSS),
+    }
+
+    for downloadable in to_download {
+        eprint!("Downloading {}... ", downloadable.filename.dimmed());
+        let contents = reqwest::get(downloadable.download_url)
+            .await?
+            .bytes()
+            .await?;
+        upgrade::write_mod_file(profile, contents, &downloadable.filename).await?;
+        println!("{}", &*TICK);
+    }
+    for installable in to_install {
+        eprint!(
+            "Installing  {}... ",
+            installable.0.to_string_lossy().dimmed()
+        );
+        copy(installable.1, profile.output_dir.join(installable.0)).await?;
+        println!("{}", &*TICK);
     }
 
     if error {